requirements.txt
pyyaml>=6.0
cryptography>=41.0.0
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
#Build from source - see https://github.com/intel/trustauthority-client-for-python
